requirements.txt
aiohttp-cors==0.7.0
aiohttp==3.7.4.post0; python_version >= "3.6"
aioredis==1.3.1
antlr4-python3-runtime==4.8; python_version >= "3.6"
async-timeout==3.0.1; python_full_version >= "3.5.3" and python_version >= "3.6"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
blessings==1.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
cachetools==4.2.2; python_version >= "3.5" and python_version < "4.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
certifi==2021.5.30; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
charset-normalizer==2.0.3; python_full_version >= "3.6.0" and python_version >= "3"
click==8.0.1; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
colorful==0.5.4
filelock==3.0.12
google-api-core==1.31.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
google-auth==1.34.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
googleapis-common-protos==1.53.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
gpustat==0.6.0
grpcio==1.39.0
hiredis==2.0.0; python_version >= "3.6"
hydra-core==1.1.0
idna==3.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
importlib-resources==5.2.1; python_version >= "3.6" and python_version < "3.9"
jsonschema==3.2.0
mordred==1.2.0
msgpack==1.0.2
multidict==5.1.0; python_version >= "3.6"
networkx==2.6.2; python_version >= "3.7"
numpy==1.21.1; python_version >= "3.7" and python_full_version >= "3.7.1"
nvidia-ml-py3==7.352.0
omegaconf==2.1.0; python_version >= "3.6"
opencensus-context==0.1.2
opencensus==0.7.13
packaging==21.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pandas==1.3.1; python_full_version >= "3.7.1"
prometheus-client==0.11.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
protobuf==3.17.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
psutil==5.8.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
py-spy==0.3.7
pyarrow==4.0.1; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.5" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_version >= "3.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and python_full_version >= "3.6.0"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyrsistent==0.18.0; python_version >= "3.6"
python-dateutil==2.8.2; python_full_version >= "3.7.1"
pytz==2021.1; python_full_version >= "3.7.1"
pyyaml==5.4.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
ray==1.2.0
redis==3.5.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
requests==2.26.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
rsa==4.7.2; python_version >= "3.5" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
six==1.16.0; python_full_version >= "3.7.1"
typing-extensions==3.10.0.0; python_version >= "3.6"
urllib3==1.26.6; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
yarl==1.6.3; python_version >= "3.6"
zipp==3.5.0; python_version >= "3.6" and python_version < "3.9"
